fix importing items when it contains id
